+++ examples/resources/meraki_switch_link_aggregations/import.sh
@@ -1,0 +1,1 @@
+terraform import meraki_switch_link_aggregations.example "<organization_id>,<network_id>"